requirements.txt
Flask
Flask-Session
mysql-connector-python
requests
